ctz/pyproject.toml
[tool.poetry]
name = "ctz"
version = "0.1.0"
description = ""
authors = ["BlurwareAdmin <admin@blurware.com>"]

[tool.poetry.scripts]
ctz = "ctz.app:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.4.1"}
emoji = "^1.7.0"
azure-identity = "^1.10.0"
azure-mgmt-resource = "^21.1.0"
azure-mgmt-storage = "^20.0.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
